fixed division, trying decimal and clewar after equals

+++ Session-09/CalculatorLib/Division.cs
@@ -9,7 +9,7 @@ namespace CalculatorLib {
         public double Do(string? x, string? y) {
             double ret = 0;
             if (x != null && y != null) {
-                ret = Convert.ToInt32(x) / Convert.ToInt32(y);
+                ret = (double)Convert.ToInt32(x) / Convert.ToInt32(y);
             }
             return ret;
         }
